custom hook for studentstats created
